server/requirements.txt
six==1.15
Django==3.0.5
graphene-django>=2.0

django-extensions==3.0.4

djoser==2.0.3
# required by djoser
django-rest-framework==0.1.0
graphene-graphiql-explorer==0.0.1

django-cors-headers==3.4.0
mock==4.0.2
django-filter==2.3.0
django-debug-toolbar==2.2

psycopg2==2.8
graphene-django-extras==0.4.9
